Peeling apart a couple files. Other misc cleanup
