+++ Src/CSharpier/DocPrinter/DocPrinter.cs
@@ -1,182 +1,25 @@
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Text;
 using CSharpier.DocTypes;
-using Newtonsoft.Json.Serialization;
+using CSharpier.Utilities;
 
 namespace CSharpier.DocPrinter
 {
-    // a big chunk of the code in here is ported from prettier. The names and layout of the file were
-    // kept consistent with how they looked in prettier because not everything
-    // was ported over and porting over more code would be easier if this file looked basically the same
-    // taken from prettier 2.2.1 or so
     public static class DocPrinter
     {
-        [ThreadStatic]
-        private static Dictionary<string, PrintMode>? groupModeMap;
-
-        private static bool Fits(
-            PrintCommand nextCommand,
-            Stack<PrintCommand> remainingCommands,
-            int remainingWidth,
-            PrinterOptions printerOptions,
-            bool mustBeFlat = false
-        ) {
-            var returnFalseIfMoreStringsFound = false;
-            var newCommands = new Stack<PrintCommand>();
-            newCommands.Push(nextCommand);
-
-            void Push(Doc doc, PrintMode printMode, Indent indent)
-            {
-                newCommands.Push(new PrintCommand(indent, printMode, doc));
-            }
-
-            var output = new StringBuilder();
-
-            for (var x = 0; x < remainingCommands.Count || newCommands.Count > 0;)
-            {
-                if (remainingWidth < 0)
-                {
-                    return false;
-                }
-
-                PrintCommand command;
-                if (newCommands.Count > 0)
-                {
-                    command = newCommands.Pop();
-                }
-                else
-                {
-                    command = remainingCommands.ElementAt(x);
-                    x++;
-                }
-
-                var (currentIndent, currentMode, currentDoc) = command;
-
-                if (currentDoc is StringDoc stringDoc)
-                {
-                    if (stringDoc.Value == null)
-                    {
-                        continue;
-                    }
-
-                    if (returnFalseIfMoreStringsFound)
-                    {
-                        return false;
-                    }
-
-                    output.Append(stringDoc.Value);
-                    remainingWidth -= GetStringWidth(stringDoc.Value);
-                }
-                else if (currentDoc != Doc.Null)
-                {
-                    switch (currentDoc)
-                    {
-                        case LeadingComment:
-                        case TrailingComment:
-                            if (output.Length > 0)
-                            {
-                                returnFalseIfMoreStringsFound = true;
-                            }
-                            break;
-                        case Concat concat:
-                            for (var i = concat.Contents.Count - 1; i >= 0; i--)
-                            {
-                                Push(concat.Contents[i], currentMode, currentIndent);
-                            }
-                            break;
-                        case IndentDoc indent:
-                            Push(
-                                indent.Contents,
-                                currentMode,
-                                IndentBuilder.Make(currentIndent, printerOptions)
-                            );
-                            break;
-                        case Trim:
-                            remainingWidth += TrimOutput(output);
-                            break;
-                        case Group group:
-                            if (mustBeFlat && group.Break)
-                            {
-                                return false;
-                            }
-
-                            var groupMode = group.Break ? PrintMode.MODE_BREAK : currentMode;
-
-                            Push(group.Contents, groupMode, currentIndent);
-
-                            if (group.GroupId != null)
-                            {
-                                groupModeMap![group.GroupId] = groupMode;
-                            }
-                            break;
-                        case IfBreak ifBreak:
-                            var ifBreakMode = ifBreak.GroupId != null
-                            && groupModeMap!.ContainsKey(ifBreak.GroupId)
-                                ? groupModeMap[ifBreak.GroupId]
-                                : currentMode;
-
-                            var contents = ifBreakMode == PrintMode.MODE_BREAK
-                                ? ifBreak.BreakContents
-                                : ifBreak.FlatContents;
-
-                            Push(contents, currentMode, currentIndent);
-                            break;
-                        case LineDoc line:
-                            switch (currentMode)
-                            {
-                                case PrintMode.MODE_FLAT:
-                                case PrintMode.MODE_FORCEFLAT:
-                                    if (currentDoc is HardLine { SkipBreakIfFirstInGroup: true })
-                                    {
-                                        returnFalseIfMoreStringsFound = false;
-                                    }
-                                    else if (line.Type == LineDoc.LineType.Hard)
-                                    {
-                                        return true;
-                                    }
-
-                                    if (line.Type != LineDoc.LineType.Soft)
-                                    {
-                                        output.Append(' ');
-
-                                        remainingWidth -= 1;
-                                    }
-                                    break;
-                                case PrintMode.MODE_BREAK:
-                                    return true;
-                            }
-                            break;
-                        case ForceFlat flat:
-                            Push(flat.Contents, currentMode, currentIndent);
-                            break;
-                        case BreakParent:
-                            break;
-                        default:
-                            throw new Exception("Can't handle " + currentDoc.GetType());
-                    }
-                }
-            }
-
-            return remainingWidth > 0;
-        }
-
         public static string Print(Doc document, PrinterOptions printerOptions, string endOfLine)
         {
-            groupModeMap = new Dictionary<string, PrintMode>();
-
             PropagateBreaks.RunOn(document);
 
-            var allowedWidth = printerOptions.Width;
-            var currentWidth = 0;
-
             var remainingCommands = new Stack<PrintCommand>();
-
             remainingCommands.Push(
-                new PrintCommand(IndentBuilder.MakeRoot(), PrintMode.MODE_BREAK, document)
+                new PrintCommand(IndentBuilder.MakeRoot(), PrintMode.Break, document)
             );
 
+            var groupModeMap = new Dictionary<string, PrintMode>();
+            var allowedWidth = printerOptions.Width;
+            var currentWidth = 0;
             var output = new StringBuilder();
             var shouldRemeasure = false;
             var newLineNextStringValue = false;
@@ -188,12 +31,12 @@ namespace CSharpier.DocPrinter
             }
             while (remainingCommands.Count > 0)
             {
-                var command = remainingCommands.Pop();
-                if (command.Doc == Doc.Null)
+                var (currentIndent, currentMode, currentDoc) = remainingCommands.Pop();
+                if (currentDoc == Doc.Null)
                 {
                     continue;
                 }
-                switch (command.Doc)
+                switch (currentDoc)
                 {
                     case StringDoc stringDoc:
                         if (string.IsNullOrEmpty(stringDoc.Value))
@@ -211,69 +54,70 @@ namespace CSharpier.DocPrinter
 
                         if (newLineNextStringValue)
                         {
-                            TrimOutput(output);
-                            output.Append(endOfLine).Append(command.Indent.Value);
-                            currentWidth = command.Indent.Length;
+                            output.TrimTrailingWhitespace();
+                            output.Append(endOfLine).Append(currentIndent.Value);
+                            currentWidth = currentIndent.Length;
                             newLineNextStringValue = false;
                         }
                         output.Append(stringDoc.Value);
-                        currentWidth += GetStringWidth(stringDoc.Value);
+                        currentWidth += stringDoc.Value.GetPrintedWidth();
                         break;
                     case Concat concat:
                         for (var x = concat.Contents.Count - 1; x >= 0; x--)
                         {
-                            Add(concat.Contents[x], command.Mode, command.Indent);
+                            Add(concat.Contents[x], currentMode, currentIndent);
                         }
                         break;
                     case IndentDoc indentBuilder:
                         Add(
                             indentBuilder.Contents,
-                            command.Mode,
-                            IndentBuilder.Make(command.Indent, printerOptions)
+                            currentMode,
+                            IndentBuilder.Make(currentIndent, printerOptions)
                         );
                         break;
                     case Trim:
-                        currentWidth -= TrimOutput(output);
+                        currentWidth -= output.TrimTrailingWhitespace();
                         newLineNextStringValue = false;
                         break;
                     case Group group:
-                        switch (command.Mode)
+                        switch (currentMode)
                         {
-                            case PrintMode.MODE_FLAT:
-                            case PrintMode.MODE_FORCEFLAT:
+                            case PrintMode.Flat:
+                            case PrintMode.Forceflat:
                                 if (!shouldRemeasure)
                                 {
                                     Add(
                                         group.Contents,
-                                        group.Break ? PrintMode.MODE_BREAK : PrintMode.MODE_FLAT,
-                                        command.Indent
+                                        group.Break ? PrintMode.Break : PrintMode.Flat,
+                                        currentIndent
                                     );
                                     break;
                                 }
 
-                                goto case PrintMode.MODE_BREAK;
-                            case PrintMode.MODE_BREAK:
+                                goto case PrintMode.Break;
+                            case PrintMode.Break:
                                 shouldRemeasure = false;
                                 var next = new PrintCommand(
-                                    command.Indent,
-                                    PrintMode.MODE_FLAT,
+                                    currentIndent,
+                                    PrintMode.Flat,
                                     group.Contents
                                 );
 
                                 if (
                                     !group.Break
-                                    && Fits(
+                                    && DocFitter.Fits(
                                         next,
                                         remainingCommands,
                                         allowedWidth - currentWidth,
-                                        printerOptions
+                                        printerOptions,
+                                        groupModeMap
                                     )
                                 ) {
                                     remainingCommands.Push(next);
                                 }
                                 else
                                 {
-                                    Add(group.Contents, PrintMode.MODE_BREAK, command.Indent);
+                                    Add(group.Contents, PrintMode.Break, currentIndent);
                                 }
                                 break;
                         }
@@ -287,17 +131,17 @@ namespace CSharpier.DocPrinter
                         var groupMode = ifBreak.GroupId != null
                         && groupModeMap.ContainsKey(ifBreak.GroupId)
                             ? groupModeMap[ifBreak.GroupId]
-                            : command.Mode;
-                        var contents = groupMode == PrintMode.MODE_BREAK
+                            : currentMode;
+                        var contents = groupMode == PrintMode.Break
                             ? ifBreak.BreakContents
                             : ifBreak.FlatContents;
-                        Add(contents, command.Mode, command.Indent);
+                        Add(contents, currentMode, currentIndent);
                         break;
                     case LineDoc line:
-                        switch (command.Mode)
+                        switch (currentMode)
                         {
-                            case PrintMode.MODE_FLAT:
-                            case PrintMode.MODE_FORCEFLAT:
+                            case PrintMode.Flat:
+                            case PrintMode.Forceflat:
                                 if (line.Type == LineDoc.LineType.Soft)
                                 {
                                     break;
@@ -313,12 +157,12 @@ namespace CSharpier.DocPrinter
                                 // group that no matter what, it needs to remeasure because the previous measurement didn't accurately
                                 // capture the entire expression (this is necessary for nested groups)
                                 shouldRemeasure = true;
-                                goto case PrintMode.MODE_BREAK;
-                            case PrintMode.MODE_BREAK:
+                                goto case PrintMode.Break;
+                            case PrintMode.Break:
                                 if (
                                     line.Squash
                                     && output.Length > 0
-                                    && EndsWithNewLineAndWhitespace(output)
+                                    && output.EndsWithNewLineAndWhitespace()
                                 ) {
                                     break;
                                 }
@@ -335,9 +179,9 @@ namespace CSharpier.DocPrinter
                                 {
                                     if (!skipNextNewLine || !newLineNextStringValue)
                                     {
-                                        TrimOutput(output);
-                                        output.Append(endOfLine).Append(command.Indent.Value);
-                                        currentWidth = command.Indent.Length;
+                                        output.TrimTrailingWhitespace();
+                                        output.Append(endOfLine).Append(currentIndent.Value);
+                                        currentWidth = currentIndent.Length;
                                     }
 
                                     if (skipNextNewLine)
@@ -351,29 +195,29 @@ namespace CSharpier.DocPrinter
                     case BreakParent:
                         break;
                     case LeadingComment leadingComment:
-                        TrimOutput(output);
+                        output.TrimTrailingWhitespace();
                         if ((output.Length != 0 && output[^1] != '\n') || newLineNextStringValue)
                         {
                             output.Append(endOfLine);
                         }
 
-                        output.Append(command.Indent.Value).Append(leadingComment.Comment);
-                        currentWidth = command.Indent.Length;
+                        output.Append(currentIndent.Value).Append(leadingComment.Comment);
+                        currentWidth = currentIndent.Length;
                         newLineNextStringValue = false;
                         skipNextNewLine = false;
                         break;
                     case TrailingComment trailingComment:
-                        TrimOutput(output);
+                        output.TrimTrailingWhitespace();
                         output.Append(' ').Append(trailingComment.Comment);
-                        currentWidth = command.Indent.Length;
+                        currentWidth = currentIndent.Length;
                         newLineNextStringValue = true;
                         skipNextNewLine = true;
                         break;
                     case ForceFlat forceFlat:
-                        Add(forceFlat.Contents, PrintMode.MODE_FLAT, command.Indent);
+                        Add(forceFlat.Contents, PrintMode.Flat, currentIndent);
                         break;
                     default:
-                        throw new Exception("didn't handle " + command.Doc);
+                        throw new Exception("didn't handle " + currentDoc);
                 }
             }
 
@@ -390,62 +234,14 @@ namespace CSharpier.DocPrinter
 
             return result;
         }
+    }
 
-        // TODO 1 in prettier this deals with unicode characters that are double width
-        private static int GetStringWidth(string value)
-        {
-            return value.Length;
-        }
+    public record PrintCommand(Indent Indent, PrintMode Mode, Doc Doc);
 
-        private static bool EndsWithNewLineAndWhitespace(StringBuilder stringBuilder)
-        {
-            for (var index = 1; index <= stringBuilder.Length; index++)
-            {
-                var next = stringBuilder[^index];
-                if (next == ' ' || next == '\t')
-                {
-                    continue;
-                }
-                else if (next == '\n')
-                {
-                    return true;
-                }
-                else
-                {
-                    return false;
-                }
-            }
-
-            return false;
-        }
-
-        private static int TrimOutput(StringBuilder stringBuilder)
-        {
-            if (stringBuilder.Length == 0)
-            {
-                return 0;
-            }
-
-            var trimmed = 0;
-            for (; trimmed < stringBuilder.Length; trimmed++)
-            {
-                if (stringBuilder[^(trimmed + 1)] != ' ' && stringBuilder[^(trimmed + 1)] != '\t')
-                {
-                    break;
-                }
-            }
-
-            stringBuilder.Length = stringBuilder.Length - trimmed;
-            return trimmed;
-        }
-
-        private record PrintCommand(Indent Indent, PrintMode Mode, Doc Doc);
-
-        private enum PrintMode
-        {
-            MODE_FLAT,
-            MODE_BREAK,
-            MODE_FORCEFLAT
-        }
+    public enum PrintMode
+    {
+        Flat,
+        Break,
+        Forceflat
     }
 }
