+++ Src/CSharpier/SyntaxPrinter/SyntaxNodePrinters/AttributeList.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using CSharpier.DocTypes;
 using CSharpier.SyntaxPrinter;
 using CSharpier.SyntaxPrinter.SyntaxNodePrinters;
+using CSharpier.Utilities;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
 namespace CSharpier.SyntaxPrinter.SyntaxNodePrinters
