+++ Src/CSharpier/SyntaxPrinter/SyntaxNodePrinters/BinaryExpression.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using CSharpier.DocTypes;
+using CSharpier.Utilities;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
