+++ Src/CSharpier/SyntaxPrinter/SyntaxNodePrinters/LabeledStatement.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using CSharpier.DocTypes;
 using CSharpier.SyntaxPrinter;
+using CSharpier.Utilities;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
 namespace CSharpier.SyntaxPrinter.SyntaxNodePrinters
