+++ Src/CSharpier/SyntaxPrinter/SyntaxNodePrinters/NamespaceDeclaration.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.Linq;
 using CSharpier.DocTypes;
 using CSharpier.SyntaxPrinter;
+using CSharpier.Utilities;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
 namespace CSharpier.SyntaxPrinter.SyntaxNodePrinters
