+++ Src/CSharpier/SyntaxPrinter/SyntaxNodePrinters/QueryBody.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using CSharpier.DocTypes;
+using CSharpier.Utilities;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
 namespace CSharpier.SyntaxPrinter.SyntaxNodePrinters
