+++ Src/CSharpier/Utilities/DirectoryExtensions.cs
@@ -1,0 +1,40 @@
+using System;
+using System.IO;
+
+namespace CSharpier.Utilities
+{
+    public static class DirectoryExtensions
+    {
+        public static void EnsureDirectoryExists(this FileInfo fileInfo)
+        {
+            fileInfo.Directory?.EnsureExists();
+        }
+
+        public static void EnsureExists(this DirectoryInfo directoryInfo)
+        {
+            if (directoryInfo.Name.EndsWith("$"))
+            {
+                return;
+            }
+
+            directoryInfo.Parent?.EnsureExists();
+
+            if (directoryInfo.Exists)
+            {
+                return;
+            }
+
+            try
+            {
+                directoryInfo.Create();
+            }
+            catch (Exception ex)
+            {
+                throw new Exception(
+                    "It was not possible to create the path " + directoryInfo.FullName,
+                    ex
+                );
+            }
+        }
+    }
+}
